Changed datatype of geo-data.

+++ src/main/java/group9/employee_management/web/dto/WorkSessionDTO.java
@@ -15,8 +15,8 @@ public class WorkSessionDTO {
     private boolean available;
     private boolean onSite;
     private String userName;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
 
     public WorkSessionDTO(){}
 
@@ -101,19 +101,19 @@ public class WorkSessionDTO {
         this.userName = userName;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 }
